+++ src/sql/index_statuses.sql
@@ -1,7 +1,0 @@
-DROP TABLE IF EXISTS index_statuses;
-CREATE TABLE index_statuses (
-    id     BIGINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    status VARCHAR(250) NOT NULL
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
-
-CREATE UNIQUE INDEX status on index_statuses (status);